+++ packages/observer/src/index.ts
@@ -2,10 +2,12 @@ export { lazyloadManager as IntersectionObserverManager } from "./base/Intersect
 export { default as IntersectionLoad } from "./components/IntersectionLoad";
 export { useBoundingClientRect } from "./hooks/useBoundingClientRect";
 export { useElementPosition } from "./hooks/useElementPosition";
+export { useElementPositionRef } from "./hooks/useElementPositionRef";
+export { useElementDetector } from "./hooks/useElementDetector";
 export { useIntersectionObserver } from "./hooks/useIntersectionObserver";
 export { useIntersectionRatio } from "./hooks/useIntersectionRatio";
 export { useInViewport } from "./hooks/useInViewport";
-export { useIsCeiling } from "./hooks/useIsCeiling";
+export { useIsMounted } from "./hooks/useIsMounted";
 export { useOneOffVisibility } from "./hooks/useOneOffVisibility";
 export { useScrollDirection } from "./hooks/useScrollDirection";
 export * from "./types";